fix: rename files to comply with google code conventions, refactor camera class
