+++ src/logger.cc
@@ -13,10 +13,10 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program. If not, see <https://www.gnu.org/licenses/>.
  *  
- *  \file Logger.cpp
+ *  \file logger.cc
  */
 
-#include "Logger.h"
+#include "logger.h"
 
 Logger::Logger() {
     // Constructor
