+++ src/logger.h
@@ -13,7 +13,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program. If not, see <https://www.gnu.org/licenses/>.
  *  
- *  \file Logger.h
+ *  \file logger.h
  */
 
 #ifndef LOGGER_H
